Organize the repo into connectors and providers. (#5)
